Conflicts:
	src/cs236369/hw5/dal/Utils.java

+++ src/cs236369/hw5/Registration.java
@@ -52,6 +52,12 @@ public class Registration extends HttpServlet {
 				response.sendRedirect("registration.jsp");
 				return;
 			}
+			stud = Student.getByUsername(stud.getStringField("username"));
+			if (stud == null) {
+				Utils.setSessionMessage(request.getSession(true), "Some error accrued, please try again");
+				response.sendRedirect("registration.jsp");
+				return;
+			}
 			if (stud.hasDuplicate()) {
 				stud.delete();
 				Utils.setSessionMessage(request.getSession(true), "Username is already taken");
